Separate test cases for the logs and results APIs of activation
Split the existing WskActivationTests into two tests, one for logs and
one for the rest of the activation apis (currently only results). This
enables a clean exclusion of the logs portion of the tests when using
the KubernetesContainerFactory (which is not able to distinguish the
stdout and stderr streams due to deficiencies in how container logs
are exposed through the Kubernetes API).
