+++ tests/src/test/scala/system/basic/WskActivationTests.scala
@@ -21,9 +21,7 @@ import common.rest.WskRestOperations
 import org.junit.runner.RunWith
 import org.scalatest.junit.JUnitRunner
 import common._
-import org.apache.openwhisk.utils.retry
 
-import scala.concurrent.duration._
 import spray.json._
 import spray.json.DefaultJsonProtocol._
 
@@ -34,30 +32,6 @@ class WskActivationTests extends TestHelpers with WskTestHelpers with WskActorSy
   val wsk: WskOperations = new WskRestOperations
 
   behavior of "Whisk activations"
-
-  it should "fetch logs using activation logs API" in withAssetCleaner(wskprops) { (wp, assetHelper) =>
-    val name = "logFetch"
-    val logFormat = "\\d+-\\d+-\\d+T\\d+:\\d+:\\d+.\\d+Z\\s+%s: %s"
-
-    assetHelper.withCleaner(wsk.action, name) { (action, _) =>
-      action.create(name, Some(TestUtils.getTestActionFilename("log.js")))
-    }
-
-    val run = wsk.action.invoke(name)
-
-    // Even though the activation was blocking, the activation itself might not have appeared in the database.
-    withActivation(wsk.activation, run) { activation =>
-      // Needs to be retried because there might be an SPI being plugged in which is handling logs not consistent with
-      // the database where the activation itself comes from (activation in CouchDB, logs in Elasticsearch for
-      // example).
-      retry({
-        val logs = wsk.activation.logs(Some(activation.activationId)).stdout
-
-        logs should include regex logFormat.format("stdout", "this is stdout")
-        logs should include regex logFormat.format("stderr", "this is stderr")
-      }, 60 * 5, Some(1.second)) // retry for 5 minutes
-    }
-  }
 
   it should "fetch result using activation result API" in withAssetCleaner(wskprops) { (wp, assetHelper) =>
     val name = "hello"
